We supported retrying a test when the test ends with over quota exception.
Update issue 7

We supported retrying a test when the test ends with over quota exception.
Now the parameters(enabled flag, max retry count, retry interval) are embedded in the code,
but they will be separated to the configuration file. This is registered in issue 18.
After supporting configuration file, we will do the refactoring.